Added wings effect and property expression Fixed unregister all Fixed particle types not changing Fixed several other small bugs Rotation expressions now take degrees for input

+++ src/main/java/me/sashie/skdragon/skript/expressions/uncommon/ExprEffectExtra.java
@@ -59,7 +59,7 @@ public class ExprEffectExtra extends CustomArrayPropertyExpression<Number> {
 	@Override
 	public void setPropertyValue(EffectData effect, int propertyNumber, Number value) {
 		if (effect instanceof IExtra) {
-			((IExtra) effect).getExtraProperty().setValue(propertyNumber, value.intValue());
+			((IExtra) effect).getExtraProperty().setValue(propertyNumber, value.floatValue());
 		}
 	}
